Import modules instead of require
Using `require` for dynamic import only works in a CJS context, and fails in an ESM context. We use rollup to output both an ESM and a CJS builds, which appropriately transforms top-level imports into `require` when transpiling the TS code. However, `require` calls used in TS code aren't transformed in the ESM output, resulting in the ESM build being unusable outside of Node 18 (where no `require` is necessary).

Using a top-level import will result in the dependency modules to be loaded even in contexts where they aren't required. However, doing a dynamic import is only supported for ESM-only builds, which isn't our case. Since the target is here server-side code, the bundle size isn't as critical, and this deosn't affect the separate browser module.